fix:멤버 리스트 페이지에서 사용하는 컴포넌트 features>member > list로 이동
