+++ src/features/member/list/GenerateInfo.tsx
