+++ src/features/member/list/MemberDetailPopup.tsx
@@ -1,14 +1,14 @@
 import { FC, useState } from 'react';
 import styled from 'styled-components';
 
-import MemberActivityInfo from '@compnents/members/MemberActivityInfo';
-import MemberBasicInfo from '@compnents/members/MemberBasicInfo';
-import MemberDetailHeader from '@compnents/members/MemberDetailHeader';
-import MemberForm from '@compnents/members/MemberForm';
+import MemberActivityInfo from 'features/member/list/MemberActivityInfo';
+import MemberBasicInfo from 'features/member/list/MemberBasicInfo';
+import MemberDetailHeader from 'features/member/list/MemberDetailHeader';
+import MemberForm from 'features/member/list/MemberForm';
 import { useUserDetailQuery } from '@queries/user/useUserDetailQuery';
 import { useMemberStore } from '@stores/memberStore';
 
-import PopupContainer from './PopupContainer';
+import PopupContainer from '../../../components/popup/PopupContainer';
 
 interface Props {
   onClose: () => void;
