+++ src/pages/admin/members/MemberList.tsx
@@ -4,9 +4,9 @@ import styled from 'styled-components';
 import Chip from '@compnents/commons/Chip';
 import FlexBox from '@compnents/commons/FlexBox';
 import Typography from '@compnents/commons/Typography';
-import MemberDetailPopup from '@compnents/popup/MemberDetailPopup';
+import MemberDetailPopup from 'features/member/list/MemberDetailPopup';
 import Pagination from '@compnents/table/Pagination';
-import StyledTable from '@compnents/table/StyledTable';
+import StyledTable from '@compnents/table/Table';
 import TableBody from '@compnents/table/TableBody';
 import TableCell from '@compnents/table/TableCell';
 import TableHead from '@compnents/table/TableHead';
