+++ src/pages/admin/notices/NoticeList.tsx
@@ -13,7 +13,7 @@ import Checkbox from '@compnents/Control/Checkbox';
 import CompletePopup from '@compnents/popup/CompletePopup';
 import ConfirmPopup from '@compnents/popup/ConfirmPopup';
 import Pagination from '@compnents/table/Pagination';
-import StyledTable from '@compnents/table/StyledTable';
+import StyledTable from '@compnents/table/Table';
 import TableBody from '@compnents/table/TableBody';
 import TableCell from '@compnents/table/TableCell';
 import TableHead from '@compnents/table/TableHead';
